split metrics_3g
